// app/types/site.ts
export interface PublicSiteConfig {
  project_name: string
  site_title: string
  site_description: string
  logo_path: string
  footer_text: string
  main_site_url: string
  login_path: string
  register_path: string
  support_path: string
  api_path: string
  support_wechat: string
  support_group_url: string
  site_url: string
  login_url: string
  register_url: string
  support_url: string
  api_base_url: string
}